+++ PyPowerStore/tests/unit_tests/test_volume_group.py
@@ -58,3 +58,17 @@ class TestVolumeGroup(TestBase):
     def test_delete_volume_group(self):
         vg = self.provisioning.delete_volume_group(self.data.vg_id1)
         self.assertIsNone(vg)
+
+    def test_clone_volume_group(self):
+        vg_clone_id = self.provisioning.clone_volume_group(self.data.vg_id1, self.data.vg_name2, None, self.data.pol_id)
+        self.assertEqual(vg_clone_id, self.data.vg_id2)
+
+    def test_refresh_volume_group(self):
+        snapshot_id = self.provisioning.refresh_volume_group(self.data.vg_id1, self.data.vg_name2,
+                                                             self.data.create_snapshot, self.data.backup_snapshot_profile)
+        self.assertEqual(snapshot_id, self.data.snapshot_id)
+
+    def test_restore_volume_group(self):
+        snapshot_id = self.provisioning.restore_volume_group(self.data.vg_id1, self.data.vg_name2,
+                                                             self.data.create_snapshot, self.data.backup_snapshot_profile)
+        self.assertEqual(snapshot_id, self.data.snapshot_id)